[package] merge test_torchscript into test_package_script
These shouldn't be separate files, they test the same thing

[ghstack-poisoned]
